pyproject.toml
[tool.poetry]
name = "physiopedia"
version = "0.1.0"
description = ""
authors = ["Andrew Phan <aphan678@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.2.14"
langchain-openai = "^0.1.22"
python-dotenv = "^1.0.1"
streamlit = "^1.37.1"
langchain-community = "^0.2.12"
langchain-groq = "^0.1.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
